requirements.txt
decorator==4.0.10
gitdb2==2.0.0
GitPython==2.1.0
nose==1.3.7
numpy==1.11.2
pandas==0.19.1
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
smmap2==2.0.1
validators==0.11.0
